// libs/enums/facility.enum.ts
export enum FacilityType {
	BADMINTON_COURTS = 'BADMINTON_COURTS',
	SWIMMING_POOLS = 'SWIMMING_POOLS',
	GOLF_FACILITIES = 'GOLF_FACILITIES',
	SQUASH_COURTS = 'SQUASH_COURTS',
	SOCCER_FIELDS = 'SOCCER_FIELDS',
	SKATING_RINKS = 'SKATING_RINKS',
	DANCE_STUDIOS = 'DANCE_STUDIOS',
	YOGA = 'YOGA',
}

export enum AvailabilityStatus {
	AVAILABLE = 'AVAILABLE',
	RESERVED = 'RESERVED',
	DELETE = 'DELETE',
}

export enum FacilityLocation {
	SEOUL = 'SEOUL',
	BUSAN = 'BUSAN',
	INCHEON = 'INCHEON',
	DAEGU = 'DAEGU',
	GYEONGJU = 'GYEONGJU',
	GWANGJU = 'GWANGJU',
	CHONJU = 'CHONJU',
	DAEJON = 'DAEJON',
	JEJU = 'JEJU',
	POHANG = 'POHANG',
	CHANGWON = 'CHANGWON',
	JEONJU = 'JEONJU',
	ULSAN = 'ULSAN',
	SUWON = 'SUWON',
	MOKPO = 'MOKPO',
}